Implement RandomScheduler. Involves major refactoring.
There was lots of redundancy between PeekScheduler and RandomScheduler. I
pulled this redundancy out into a mix-in trait: ExternalEventInjector.

There was also redundanct between PeekScheduler, FairScheduler, and
RandomScheduler. I pulled this out into an abstract superclass:
AbstractScheduler. It basically just tracks the names of actors in the system.

Slight change to the FailureDetector interface: take an anonymous function
rather than an entire scheduler.