Properly kill actors, without getting into deadlock. Issue:
Suppose that we just delivered a message to an actor that we're not about to
kill. Aspectj's "after(): ActorCell.invoke" will kick in and have us dispatch
a new message (invoke advanceTrace) where we will now kill the actor.

This causes a deadlock. The issue is that ActorCell.invoke is still *within*
the actor's mailbox. The actor's mailbox state has not yet been set to "Idle"
at this point, so it cannot process any system messages. Akka uses system
messages (`Terminate`) to kill the actor, but it can't actually kill the actor
because the mailbox isn't set as ready to accept messages.

The solution here is to *wait* until after the mailbox has been set to "Idle"
state, then start dispatch again so that we can go ahead and kill it.